refactor: Вынести компоненты header блока из папки ui
